feat(valid-package-definition): rename rule from valid-package-def
This change deprecates the `valid-package-def` rule in favor of a new `valid-package-definition` rule, providing the same behavior.  The deprecated version of the rule will be removed in the next major release.
