+++ src/rules/valid-package-definition.ts
